+++ CHANGELOG.rst
@@ -224,6 +224,10 @@ v0.5
     - ablility to filter flat chain to be within specific interval of parameters using AnalyticsTask.filter_chain
       function. This method is suitable for examining multiple solutions.
     - ability to evaluate R^2 for set of model parameters using function AnalyticsTask.coefficient_of_determination.
+    - adding ability to load from json in "radius"-based format that describes the size of the star with
+      `equivalent_radius` instead of polar gravity `polar_log_g` in "standard" format
+    - BinarySystem and SingleSystem now contain a function build_container that builds a complete model of a system at
+      given photometric `phase` or observational `time`.
 
 
 **Fixes**